+++ src/ngm-app.ts
@@ -6,7 +6,6 @@ import './elements/ngm-full-screen-view';
 import './elements/ngm-object-information';
 import './elements/ngm-feature-height';
 import './elements/ngm-auth';
-import './elements/ngm-drop-files';
 import './elements/ngm-tracking-consent';
 import './elements/ngm-camera-information.js';
 import './elements/ngm-nav-tools';
@@ -30,7 +29,6 @@ import QueryManager from './query/QueryManager';
 
 import {initAnalytics} from './analytics.js';
 import {initSentry} from './sentry.js';
-import {showWarning} from './message';
 import MainStore from './store/main';
 import SlicerStore from './store/slicer';
 import {classMap} from 'lit-html/directives/class-map.js';
@@ -39,8 +37,6 @@ import MapChooser from './MapChooser';
 import {NgmLayerLegendContainer} from './elements/ngm-layer-legend-container';
 import {NgmSlowLoading} from './elements/ngm-slow-loading';
 import {NgmSlicer} from './elements/slicer/ngm-slicer';
-import {NgmAreaOfInterestDrawer} from './toolbox/AreaOfInterestDrawer';
-import LayersUpload from './layers/ngm-layers-upload';
 import {NgmLoadingMask} from './elements/ngm-loading-mask';
 import {Viewer} from 'cesium';
 
@@ -129,29 +125,6 @@ export class NgmApp extends LitElementI18n {
     setupSearch(viewer, this.querySelector('ga-search')!, sideBar);
   }
 
-
-  /**
-   * @param file
-   * @param {'toolbox'|'model'} type
-   */
-  onFileDrop(file, type) {
-    if (type === 'toolbox') {
-      const aoi: NgmAreaOfInterestDrawer = this.querySelector('ngm-aoi-drawer')!;
-      if (file.name.toLowerCase().endsWith('.kml')) {
-        aoi.uploadKml(file);
-      } else if (file.name.toLowerCase().endsWith('.gpx')) {
-        aoi.uploadGpx(file);
-      }
-    } else if (type === 'model') {
-      if (file.name.toLowerCase().endsWith('.kml')) {
-        const kmlUpload: LayersUpload = this.querySelector('ngm-layers-upload')!;
-        kmlUpload.uploadKml(file);
-      } else {
-        showWarning(i18next.t('dtd_file_not_kml'));
-      }
-    }
-  }
-
   /**
    * @param {import ('cesium').Viewer} viewer
    */
@@ -188,8 +161,6 @@ export class NgmApp extends LitElementI18n {
         unlisten();
       }
     });
-
-
   }
 
   firstUpdated() {
@@ -265,8 +236,6 @@ export class NgmApp extends LitElementI18n {
         <ngm-camera-information .viewer="${this.viewer}"></ngm-camera-information>
       </header>
       <main>
-        <ngm-drop-files @filedrop="${event => this.onFileDrop(event.detail.file, event.detail.type)}"
-                        .target="${document.body}"></ngm-drop-files>
         <ngm-loading-mask></ngm-loading-mask>
         <ngm-side-bar
           .queryManager=${this.queryManager}
