+++ src/style/index.css
@@ -39,7 +39,6 @@
 @import 'ngm-camera-information.css';
 @import 'ngm-position-edit.css';
 @import 'ngm-map-chooser.css';
-@import 'ngm-drop-files.css';
 @import 'ngm-tracking-consent.css';
 @import 'ngm-layer-legend.css';
 @import 'icons.css';
